// src/types.ts
export type SpecialMenu = {
  name: string
  cards: string[]
  kick: number
  speed: number
  stamina: number
  technique: number
  physical: number
  jump: number
  mental: number
  total: number
  fatigue: number
  offenseQuality: number
  support: number
  triangle: number
  loseMark: number
  overlap: number
  diagonalRun: number
  spaceRun: number
  goalFront: number
  defenseQuality: number
  zoneMarking: number
  manMarking: number
  pressing: number
  shootCut: number
  intercept: number
}